Changed the scope in which examples are executed.
Examples ('it'-blocks) are now executed on the current ExampleGroup
instance using XOTcls 'eval' method, instead of the homebrewn
'instance_eval' method.

This has the effect that variables that are defined inside a
'before each' block are available in the examples, without having
to bring them into the scope using 'my instvar'.

This reduces noise inside the examples and increases the readability.

Also, examples are no longer executed in the ::Spec namespace, but
instead run in the ::xotcl::fakeNS namespace, but that should not be
of any concern.